Fixed domain-model mix-up after conversion to Spring Boot.
