+++ cart-service/src/main/java/com/redhat/coolstore/model/Product.java
@@ -1,54 +1,80 @@
 package com.redhat.coolstore.model;
 
-import java.io.Serializable;
+/**
+ * This class was automatically generated by the data modeler tool.
+ */
+@org.kie.api.definition.type.Label("Product")
+public class Product implements java.io.Serializable
+{
 
-public class Product implements Serializable {
+   static final long serialVersionUID = 1L;
 
-	private static final long serialVersionUID = -7304814269819778382L;
-	private String itemId;
-	private String name;
-	private String desc;
-	private double price;
-	
-	public Product() {
-		
-	}
-	
-	public Product(String itemId, String name, String desc, double price) {
-		super();
-		this.itemId = itemId;
-		this.name = name;
-		this.desc = desc;
-		this.price = price;
-	}
-	public String getItemId() {
-		return itemId;
-	}
-	public void setItemId(String itemId) {
-		this.itemId = itemId;
-	}
-	public String getName() {
-		return name;
-	}
-	public void setName(String name) {
-		this.name = name;
-	}
-	public String getDesc() {
-		return desc;
-	}
-	public void setDesc(String desc) {
-		this.desc = desc;
-	}
-	public double getPrice() {
-		return price;
-	}
-	public void setPrice(double price) {
-		this.price = price;
-	}
+   @org.kie.api.definition.type.Label("Description")
+   @org.kie.api.definition.type.Position(2)
+   private java.lang.String desc;
 
-	@Override
-	public String toString() {
-		return "Product [itemId=" + itemId + ", name=" + name + ", desc="
-				+ desc + ", price=" + price + "]";
-	}
+   @org.kie.api.definition.type.Label("Item ID")
+   @org.kie.api.definition.type.Position(0)
+   private java.lang.String itemId;
+
+   @org.kie.api.definition.type.Label("Name")
+   @org.kie.api.definition.type.Position(1)
+   private java.lang.String name;
+
+   @org.kie.api.definition.type.Label("Price")
+   @org.kie.api.definition.type.Position(3)
+   private double price;
+
+   public Product()
+   {
+   }
+
+   public java.lang.String getDesc()
+   {
+      return this.desc;
+   }
+
+   public void setDesc(java.lang.String desc)
+   {
+      this.desc = desc;
+   }
+
+   public java.lang.String getItemId()
+   {
+      return this.itemId;
+   }
+
+   public void setItemId(java.lang.String itemId)
+   {
+      this.itemId = itemId;
+   }
+
+   public java.lang.String getName()
+   {
+      return this.name;
+   }
+
+   public void setName(java.lang.String name)
+   {
+      this.name = name;
+   }
+
+   public double getPrice()
+   {
+      return this.price;
+   }
+
+   public void setPrice(double price)
+   {
+      this.price = price;
+   }
+
+   public Product(java.lang.String desc, java.lang.String itemId,
+         java.lang.String name, double price)
+   {
+      this.desc = desc;
+      this.itemId = itemId;
+      this.name = name;
+      this.price = price;
+   }
 }
