+++ cart-service/src/main/java/com/redhat/coolstore/model/PromoEvent.java
@@ -1,4 +1,4 @@
-package com.redhat.coolstore.model.kie;
+package com.redhat.coolstore.model;
 
 /**
  * This class was automatically generated by the data modeler tool.
