+++ cart-service/src/main/java/com/redhat/coolstore/model/ShoppingCart.java
@@ -1,113 +1,115 @@
 package com.redhat.coolstore.model;
 
-import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.List;
+/**
+ * This class was automatically generated by the data modeler tool.
+ */
+@org.kie.api.definition.type.Label("Shopping Cart")
+public class ShoppingCart implements java.io.Serializable
+{
 
-public class ShoppingCart implements Serializable {
+   static final long serialVersionUID = 1L;
 
-	private static final long serialVersionUID = -1108043957592113528L;
+   @org.kie.api.definition.type.Label("Cart Item Promotional Savings")
+   @org.kie.api.definition.type.Position(1)
+   private double cartItemPromoSavings;
 
-	private double cartItemTotal;
+   @org.kie.api.definition.type.Label("Cart Item Total")
+   @org.kie.api.definition.type.Position(0)
+   private double cartItemTotal;
 
-	private double cartItemPromoSavings;
-	
-	private double shippingTotal;
-	
-	private double shippingPromoSavings;
-	
-	private double cartTotal;
-			
-	private List<ShoppingCartItem> shoppingCartItemList = new ArrayList<ShoppingCartItem>();
+   @org.kie.api.definition.type.Label("Shopping Cart Total")
+   @org.kie.api.definition.type.Position(4)
+   private double cartTotal;
 
-	public ShoppingCart() {
-		
-	}
-	
-	public List<ShoppingCartItem> getShoppingCartItemList() {
-		return shoppingCartItemList;
-	}
+   @org.kie.api.definition.type.Label("Shipping Promotional Savings")
+   @org.kie.api.definition.type.Position(3)
+   private double shippingPromoSavings;
 
-	public void setShoppingCartItemList(List<ShoppingCartItem> shoppingCartItemList) {
-		this.shoppingCartItemList = shoppingCartItemList;
-	}
+   @org.kie.api.definition.type.Label("Shipping Total")
+   @org.kie.api.definition.type.Position(2)
+   private double shippingTotal;
 
-	public void resetShoppingCartItemList() {
-		shoppingCartItemList = new ArrayList<ShoppingCartItem>();
-	}
+   @org.kie.api.definition.type.Label("List of Shopping Cart Items")
+   @org.kie.api.definition.type.Position(5)
+   private java.util.List<com.redhat.coolstore.model.ShoppingCartItem> shoppingCartItemList;
 
-	public void addShoppingCartItem(ShoppingCartItem sci) {
-		
-		if ( sci != null ) {
-			
-			shoppingCartItemList.add(sci);
-			
-		}
-		
-	}
-	
-	public boolean removeShoppingCartItem(ShoppingCartItem sci) {
-		
-		boolean removed = false;
-		
-		if ( sci != null ) {
-			
-			removed = shoppingCartItemList.remove(sci);
-			
-		}
-		
-		return removed;
-		
-	}
+   public ShoppingCart()
+   {
+   }
 
-	public double getCartItemTotal() {
-		return cartItemTotal;
-	}
+   public java.util.List<com.redhat.coolstore.model.ShoppingCartItem> getShoppingCartItemList()
+   {
+      return this.shoppingCartItemList;
+   }
 
-	public void setCartItemTotal(double cartItemTotal) {
-		this.cartItemTotal = cartItemTotal;
-	}
+   public void setShoppingCartItemList(java.util.List<com.redhat.coolstore.model.ShoppingCartItem> shoppingCartItemList)
+   {
+      this.shoppingCartItemList = shoppingCartItemList;
+   }
 
-	public double getShippingTotal() {
-		return shippingTotal;
-	}
+   public double getCartItemPromoSavings()
+   {
+      return this.cartItemPromoSavings;
+   }
 
-	public void setShippingTotal(double shippingTotal) {
-		this.shippingTotal = shippingTotal;
-	}
+   public void setCartItemPromoSavings(double cartItemPromoSavings)
+   {
+      this.cartItemPromoSavings = cartItemPromoSavings;
+   }
 
-	public double getCartTotal() {
-		return cartTotal;
-	}
+   public double getCartItemTotal()
+   {
+      return this.cartItemTotal;
+   }
 
-	public void setCartTotal(double cartTotal) {
-		this.cartTotal = cartTotal;
-	}
+   public void setCartItemTotal(double cartItemTotal)
+   {
+      this.cartItemTotal = cartItemTotal;
+   }
 
-	public double getCartItemPromoSavings() {
-		return cartItemPromoSavings;
-	}
+   public double getCartTotal()
+   {
+      return this.cartTotal;
+   }
 
-	public void setCartItemPromoSavings(double cartItemPromoSavings) {
-		this.cartItemPromoSavings = cartItemPromoSavings;
-	}
+   public void setCartTotal(double cartTotal)
+   {
+      this.cartTotal = cartTotal;
+   }
 
-	public double getShippingPromoSavings() {
-		return shippingPromoSavings;
-	}
+   public double getShippingPromoSavings()
+   {
+      return this.shippingPromoSavings;
+   }
 
-	public void setShippingPromoSavings(double shippingPromoSavings) {
-		this.shippingPromoSavings = shippingPromoSavings;
-	}
+   public void setShippingPromoSavings(double shippingPromoSavings)
+   {
+      this.shippingPromoSavings = shippingPromoSavings;
+   }
 
-	@Override
-	public String toString() {
-		return "ShoppingCart [cartItemTotal=" + cartItemTotal
-				+ ", cartItemPromoSavings=" + cartItemPromoSavings
-				+ ", shippingTotal=" + shippingTotal
-				+ ", shippingPromoSavings=" + shippingPromoSavings
-				+ ", cartTotal=" + cartTotal + ", shoppingCartItemList="
-				+ shoppingCartItemList + "]";
-	}
-	
+   public double getShippingTotal()
+   {
+      return this.shippingTotal;
+   }
+
+   public void setShippingTotal(double shippingTotal)
+   {
+      this.shippingTotal = shippingTotal;
+   }
+
+   public ShoppingCart(
+         double cartItemPromoSavings,
+         double cartItemTotal,
+         double cartTotal,
+         double shippingPromoSavings,
+         double shippingTotal,
+         java.util.List<com.redhat.coolstore.model.ShoppingCartItem> shoppingCartItemList)
+   {
+      this.cartItemPromoSavings = cartItemPromoSavings;
+      this.cartItemTotal = cartItemTotal;
+      this.cartTotal = cartTotal;
+      this.shippingPromoSavings = shippingPromoSavings;
+      this.shippingTotal = shippingTotal;
+      this.shoppingCartItemList = shoppingCartItemList;
+   }
 }
