+++ cart-service/src/main/java/com/redhat/coolstore/model/kie/Promotion.java
@@ -1,4 +1,4 @@
-package com.redhat.coolstore.model;
+package com.redhat.coolstore.model.kie;
 
 public class Promotion {
