+++ cart-service/src/main/java/com/redhat/coolstore/model/kie/ShoppingCart.java
@@ -1,115 +1,113 @@
 package com.redhat.coolstore.model.kie;
 
-/**
- * This class was automatically generated by the data modeler tool.
- */
-@org.kie.api.definition.type.Label("Shopping Cart")
-public class ShoppingCart implements java.io.Serializable
-{
+import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.List;
 
-   static final long serialVersionUID = 1L;
+public class ShoppingCart implements Serializable {
 
-   @org.kie.api.definition.type.Label("Cart Item Promotional Savings")
-   @org.kie.api.definition.type.Position(1)
-   private double cartItemPromoSavings;
+	private static final long serialVersionUID = -1108043957592113528L;
 
-   @org.kie.api.definition.type.Label("Cart Item Total")
-   @org.kie.api.definition.type.Position(0)
-   private double cartItemTotal;
+	private double cartItemTotal;
 
-   @org.kie.api.definition.type.Label("Shopping Cart Total")
-   @org.kie.api.definition.type.Position(4)
-   private double cartTotal;
+	private double cartItemPromoSavings;
+	
+	private double shippingTotal;
+	
+	private double shippingPromoSavings;
+	
+	private double cartTotal;
+			
+	private List<ShoppingCartItem> shoppingCartItemList = new ArrayList<ShoppingCartItem>();
 
-   @org.kie.api.definition.type.Label("Shipping Promotional Savings")
-   @org.kie.api.definition.type.Position(3)
-   private double shippingPromoSavings;
+	public ShoppingCart() {
+		
+	}
+	
+	public List<ShoppingCartItem> getShoppingCartItemList() {
+		return shoppingCartItemList;
+	}
 
-   @org.kie.api.definition.type.Label("Shipping Total")
-   @org.kie.api.definition.type.Position(2)
-   private double shippingTotal;
+	public void setShoppingCartItemList(List<ShoppingCartItem> shoppingCartItemList) {
+		this.shoppingCartItemList = shoppingCartItemList;
+	}
 
-   @org.kie.api.definition.type.Label("List of Shopping Cart Items")
-   @org.kie.api.definition.type.Position(5)
-   private java.util.List<com.redhat.coolstore.model.kie.ShoppingCartItem> shoppingCartItemList;
+	public void resetShoppingCartItemList() {
+		shoppingCartItemList = new ArrayList<ShoppingCartItem>();
+	}
 
-   public ShoppingCart()
-   {
-   }
+	public void addShoppingCartItem(ShoppingCartItem sci) {
+		
+		if ( sci != null ) {
+			
+			shoppingCartItemList.add(sci);
+			
+		}
+		
+	}
+	
+	public boolean removeShoppingCartItem(ShoppingCartItem sci) {
+		
+		boolean removed = false;
+		
+		if ( sci != null ) {
+			
+			removed = shoppingCartItemList.remove(sci);
+			
+		}
+		
+		return removed;
+		
+	}
 
-   public java.util.List<com.redhat.coolstore.model.kie.ShoppingCartItem> getShoppingCartItemList()
-   {
-      return this.shoppingCartItemList;
-   }
+	public double getCartItemTotal() {
+		return cartItemTotal;
+	}
 
-   public void setShoppingCartItemList(java.util.List<com.redhat.coolstore.model.kie.ShoppingCartItem> shoppingCartItemList)
-   {
-      this.shoppingCartItemList = shoppingCartItemList;
-   }
+	public void setCartItemTotal(double cartItemTotal) {
+		this.cartItemTotal = cartItemTotal;
+	}
 
-   public double getCartItemPromoSavings()
-   {
-      return this.cartItemPromoSavings;
-   }
+	public double getShippingTotal() {
+		return shippingTotal;
+	}
 
-   public void setCartItemPromoSavings(double cartItemPromoSavings)
-   {
-      this.cartItemPromoSavings = cartItemPromoSavings;
-   }
+	public void setShippingTotal(double shippingTotal) {
+		this.shippingTotal = shippingTotal;
+	}
 
-   public double getCartItemTotal()
-   {
-      return this.cartItemTotal;
-   }
+	public double getCartTotal() {
+		return cartTotal;
+	}
 
-   public void setCartItemTotal(double cartItemTotal)
-   {
-      this.cartItemTotal = cartItemTotal;
-   }
+	public void setCartTotal(double cartTotal) {
+		this.cartTotal = cartTotal;
+	}
 
-   public double getCartTotal()
-   {
-      return this.cartTotal;
-   }
+	public double getCartItemPromoSavings() {
+		return cartItemPromoSavings;
+	}
 
-   public void setCartTotal(double cartTotal)
-   {
-      this.cartTotal = cartTotal;
-   }
+	public void setCartItemPromoSavings(double cartItemPromoSavings) {
+		this.cartItemPromoSavings = cartItemPromoSavings;
+	}
 
-   public double getShippingPromoSavings()
-   {
-      return this.shippingPromoSavings;
-   }
+	public double getShippingPromoSavings() {
+		return shippingPromoSavings;
+	}
 
-   public void setShippingPromoSavings(double shippingPromoSavings)
-   {
-      this.shippingPromoSavings = shippingPromoSavings;
-   }
+	public void setShippingPromoSavings(double shippingPromoSavings) {
+		this.shippingPromoSavings = shippingPromoSavings;
+	}
 
-   public double getShippingTotal()
-   {
-      return this.shippingTotal;
-   }
-
-   public void setShippingTotal(double shippingTotal)
-   {
-      this.shippingTotal = shippingTotal;
-   }
-
-   public ShoppingCart(
-         double cartItemPromoSavings,
-         double cartItemTotal,
-         double cartTotal,
-         double shippingPromoSavings,
-         double shippingTotal,
-         java.util.List<com.redhat.coolstore.model.kie.ShoppingCartItem> shoppingCartItemList)
-   {
-      this.cartItemPromoSavings = cartItemPromoSavings;
-      this.cartItemTotal = cartItemTotal;
-      this.cartTotal = cartTotal;
-      this.shippingPromoSavings = shippingPromoSavings;
-      this.shippingTotal = shippingTotal;
-      this.shoppingCartItemList = shoppingCartItemList;
-   }
+	@Override
+	public String toString() {
+		return "ShoppingCart [cartItemTotal=" + cartItemTotal
+				+ ", cartItemPromoSavings=" + cartItemPromoSavings
+				+ ", shippingTotal=" + shippingTotal
+				+ ", shippingPromoSavings=" + shippingPromoSavings
+				+ ", cartTotal=" + cartTotal + ", shoppingCartItemList="
+				+ shoppingCartItemList + "]";
+	}
+	
 }
