+++ cart-service/src/main/java/com/redhat/coolstore/model/kie/ShoppingCartItem.java
@@ -1,123 +1,58 @@
 package com.redhat.coolstore.model.kie;
 
-/**
- * This class was automatically generated by the data modeler tool.
- */
-@org.kie.api.definition.type.Label("Shopping Cart Item")
-public class ShoppingCartItem implements java.io.Serializable
-{
+import java.io.Serializable;
 
-   static final long serialVersionUID = 1L;
+public class ShoppingCartItem implements Serializable {
+	
+	private static final long serialVersionUID = 6964558044240061049L;
 
-   @org.kie.api.definition.type.Label("Item ID")
-   @org.kie.api.definition.type.Position(0)
-   private java.lang.String itemId;
+	private double price;
+	private int quantity;
+	private double promoSavings;
+	private Product product;
+	
+	public ShoppingCartItem() {
+		
+	}
+	
+	public double getPrice() {
+		return price;
+	}
 
-   @org.kie.api.definition.type.Label("Name")
-   @org.kie.api.definition.type.Position(1)
-   private java.lang.String name;
+	public void setPrice(double price) {
+		this.price = price;
+	}
 
-   @org.kie.api.definition.type.Label("Price")
-   @org.kie.api.definition.type.Position(2)
-   private double price;
+	public Product getProduct() {
+		return product;
+	}
 
-   @org.kie.api.definition.type.Label("Promotional Savings")
-   @org.kie.api.definition.type.Position(4)
-   private double promoSavings;
+	public void setProduct(Product product) {
+		this.product = product;
+	}
 
-   @org.kie.api.definition.type.Label("Quantity")
-   @org.kie.api.definition.type.Position(3)
-   private int quantity;
+	public int getQuantity() {
+		return quantity;
+	}
 
-   @org.kie.api.definition.type.Label("Shopping Cart")
-   @org.kie.api.definition.type.Position(5)
-   private com.redhat.coolstore.model.kie.ShoppingCart shoppingCart;
+	public void setQuantity(int quantity) {
+		this.quantity = quantity;
+	}
 
-   public ShoppingCartItem()
-   {
-   }
+	public double getPromoSavings() {
+		return promoSavings;
+	}
 
-   public java.lang.String getItemId()
-   {
-      return this.itemId;
-   }
+	public void setPromoSavings(double promoSavings) {
+		this.promoSavings = promoSavings;
+	}
 
-   public void setItemId(java.lang.String itemId)
-   {
-      this.itemId = itemId;
-   }
-
-   public java.lang.String getName()
-   {
-      return this.name;
-   }
-
-   public void setName(java.lang.String name)
-   {
-      this.name = name;
-   }
-
-   public com.redhat.coolstore.model.kie.ShoppingCart getShoppingCart()
-   {
-      return this.shoppingCart;
-   }
-
-   public void setShoppingCart(com.redhat.coolstore.model.kie.ShoppingCart shoppingCart)
-   {
-      this.shoppingCart = shoppingCart;
-   }
-
-   public double getPrice()
-   {
-      return this.price;
-   }
-
-   public void setPrice(double price)
-   {
-      this.price = price;
-   }
-
-   public double getPromoSavings()
-   {
-      return this.promoSavings;
-   }
-
-   public void setPromoSavings(double promoSavings)
-   {
-      this.promoSavings = promoSavings;
-   }
-
-   public int getQuantity()
-   {
-      return this.quantity;
-   }
-
-   public void setQuantity(int quantity)
-   {
-      this.quantity = quantity;
-   }
-
-   public ShoppingCartItem(java.lang.String itemId, java.lang.String name,
-         double price, double promoSavings, int quantity,
-         com.redhat.coolstore.model.kie.ShoppingCart shoppingCart)
-   {
-      this.itemId = itemId;
-      this.name = name;
-      this.price = price;
-      this.promoSavings = promoSavings;
-      this.quantity = quantity;
-      this.shoppingCart = shoppingCart;
-   }
-
-   public ShoppingCartItem(java.lang.String itemId, java.lang.String name,
-         double price, int quantity, double promoSavings,
-         com.redhat.coolstore.model.kie.ShoppingCart shoppingCart)
-   {
-      this.itemId = itemId;
-      this.name = name;
-      this.price = price;
-      this.quantity = quantity;
-      this.promoSavings = promoSavings;
-      this.shoppingCart = shoppingCart;
-   }
+	@Override
+	public String toString() {
+		return "ShoppingCartItem [price=" + price + ", quantity=" + quantity
+				+ ", promoSavings=" + promoSavings + ", product=" + product
+				+ "]";
+	}
+		
+	
 }
