+++ cart-service/src/main/java/com/redhat/coolstore/service/CatalogService.java
@@ -6,7 +6,7 @@ import org.springframework.cloud.netflix.feign.FeignClient;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
-import com.redhat.coolstore.model.Product;
+import com.redhat.coolstore.model.kie.Product;
 
 @FeignClient(name = "catalogService", url = "${CATALOG_ENDPOINT}")
 interface CatalogService {
