+++ cart-service/src/main/java/com/redhat/coolstore/service/ShoppingCartServiceImplDecisionServer.java
@@ -28,14 +28,11 @@ import org.springframework.context.annotation.Primary;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 
-import com.redhat.coolstore.model.Product;
-import com.redhat.coolstore.model.Promotion;
-import com.redhat.coolstore.model.ShoppingCart;
-import com.redhat.coolstore.model.ShoppingCartItem;
-import com.redhat.coolstore.model.kie.PromoEvent;
-
-import feign.Feign;
-import feign.jackson.JacksonDecoder;
+import com.redhat.coolstore.model.PromoEvent;
+import com.redhat.coolstore.model.kie.Product;
+import com.redhat.coolstore.model.kie.Promotion;
+import com.redhat.coolstore.model.kie.ShoppingCart;
+import com.redhat.coolstore.model.kie.ShoppingCartItem;
 
 @Service
 @Primary
@@ -44,7 +41,7 @@ public class ShoppingCartServiceImplDecisionServer implements ShoppingCartServic
 
 	private static final Logger LOGGER = LoggerFactory.getLogger(ShoppingCartServiceImplDecisionServer.class);
 
-	private static final String CATALOG_ENDPOINT = System.getenv("CATALOG_ENDPOINT");
+	//private static final String CATALOG_ENDPOINT = System.getenv("CATALOG_ENDPOINT");
 	private static final String PRICING_ENDPOINT = System.getenv("PRICING_ENDPOINT");
 	private static final String URL = PRICING_ENDPOINT + "/kie-server/services/rest/server";
 	private static final String USER = System.getenv("KIE_SERVER_USER");
@@ -106,7 +103,7 @@ public class ShoppingCartServiceImplDecisionServer implements ShoppingCartServic
 
 			if (executeResponse.getType() == ResponseType.SUCCESS) {
 				ExecutionResults results = executeResponse.getResult();
-				com.redhat.coolstore.model.kie.ShoppingCart resultSc = (com.redhat.coolstore.model.kie.ShoppingCart) results.getValue("shoppingcart");
+				com.redhat.coolstore.model.ShoppingCart resultSc = (com.redhat.coolstore.model.ShoppingCart) results.getValue("shoppingcart");
 				mapShoppingCartPricingResults(resultSc, sc);
 			} else {
 				// TODO: Some proper, micro-service type error handling here.
@@ -152,7 +149,7 @@ public class ShoppingCartServiceImplDecisionServer implements ShoppingCartServic
 		/*
 		 * Build the ShoppingCart fact from the given ShoppingCart.
 		 */
-		com.redhat.coolstore.model.kie.ShoppingCart factSc = buildShoppingCartFact(sc);
+		com.redhat.coolstore.model.ShoppingCart factSc = buildShoppingCartFact(sc);
 
 		commands.add(commandsFactory.newInsert(factSc, "shoppingcart", true, "DEFAULT"));
 
@@ -160,7 +157,7 @@ public class ShoppingCartServiceImplDecisionServer implements ShoppingCartServic
 		List<ShoppingCartItem> scItems = sc.getShoppingCartItemList();
 		for (ShoppingCartItem nextSci : scItems) {
 			// Build the ShoppingCartItem fact from the given ShoppingCartItem.
-			com.redhat.coolstore.model.kie.ShoppingCartItem factSci = buildShoppingCartItem(nextSci);
+			com.redhat.coolstore.model.ShoppingCartItem factSci = buildShoppingCartItem(nextSci);
 			factSci.setShoppingCart(factSc);
 			commands.add(commandsFactory.newInsert(factSci));
 		}
@@ -204,8 +201,8 @@ public class ShoppingCartServiceImplDecisionServer implements ShoppingCartServic
 	 *            the {@link ShoppingCart} from which to build the fact.
 	 * @return the {@link com.redhat.coolstore.ShoppingCart} fact
 	 */
-	private com.redhat.coolstore.model.kie.ShoppingCart buildShoppingCartFact(ShoppingCart sc) {
-		com.redhat.coolstore.model.kie.ShoppingCart factSc = new com.redhat.coolstore.model.kie.ShoppingCart();
+	private com.redhat.coolstore.model.ShoppingCart buildShoppingCartFact(ShoppingCart sc) {
+		com.redhat.coolstore.model.ShoppingCart factSc = new com.redhat.coolstore.model.ShoppingCart();
 		factSc.setCartItemPromoSavings(sc.getCartItemPromoSavings());
 		factSc.setCartItemTotal(sc.getCartItemTotal());
 		factSc.setCartTotal(sc.getCartTotal());
@@ -221,8 +218,8 @@ public class ShoppingCartServiceImplDecisionServer implements ShoppingCartServic
 	 *            the {@link ShoppingCartItem} from which to build the fact.
 	 * @return the {@link com.redhat.coolstore.ShoppingCartItem} fact.
 	 */
-	private com.redhat.coolstore.model.kie.ShoppingCartItem buildShoppingCartItem(ShoppingCartItem sci) {
-		com.redhat.coolstore.model.kie.ShoppingCartItem factSci = new com.redhat.coolstore.model.kie.ShoppingCartItem();
+	private com.redhat.coolstore.model.ShoppingCartItem buildShoppingCartItem(ShoppingCartItem sci) {
+		com.redhat.coolstore.model.ShoppingCartItem factSci = new com.redhat.coolstore.model.ShoppingCartItem();
 		factSci.setItemId(sci.getProduct().getItemId());
 		factSci.setName(sci.getProduct().getName());
 		factSci.setPrice(sci.getProduct().getPrice());
@@ -238,7 +235,7 @@ public class ShoppingCartServiceImplDecisionServer implements ShoppingCartServic
 	 * @param sc
 	 *            the {@link ShoppingCart} onto which we need to map the results.
 	 */
-	private void mapShoppingCartPricingResults(com.redhat.coolstore.model.kie.ShoppingCart resultSc, ShoppingCart sc) {
+	private void mapShoppingCartPricingResults(com.redhat.coolstore.model.ShoppingCart resultSc, ShoppingCart sc) {
 		sc.setCartItemPromoSavings(resultSc.getCartItemPromoSavings());
 		sc.setCartItemTotal(resultSc.getCartItemTotal());
 		sc.setShippingPromoSavings(resultSc.getShippingPromoSavings());
